+++ src/test/scala/grizzled/security.scala
@@ -1,5 +1,6 @@
+package grizzled
 
-import org.scalatest.FunSuite
+import org.scalatest.{FlatSpec, Matchers}
 import grizzled.security._
 import scala.io.Source
 import java.io.ByteArrayInputStream
@@ -7,46 +8,29 @@ import java.io.ByteArrayInputStream
 /**
  * Tests the grizzled.security functions.
  */
-class SecurityTest extends FunSuite {
+class SecuritySpec extends FlatSpec with Matchers {
   val Data = Array(
     ("sha-256", "foo") -> "2c26b46b68ffc68ff99b453c1d30413413422d706483bfa0f98a5e886266e7ae",
     ("md5", "foo")     -> "acbd18db4cc2f85cedef654fccc4a4d8"
   )
 
-  test("MessageDigest: string inputs") {
+  "MessageDigest" should "work with string inputs" in {
 
-    for (testItem <- Data) {
-      val (algorithm, str) = testItem._1
-      val expected         = testItem._2
-
-      assertResult(expected) {
-        MessageDigest(algorithm).digestString(str)
-      }
+    for (((algorithm, str), expected) <- Data) {
+      MessageDigest(algorithm).digestString(str) shouldBe expected
     }
   }
 
-  test("MessageDigest: source inputs") {
-
-    for (testItem <- Data) {
-      val (algorithm, str) = testItem._1
-      val expected         = testItem._2
-
-      assertResult(expected) {
-        MessageDigest(algorithm).digestString(Source.fromString(str))
-      }
+  it should "work with Source inputs" in {
+    for (((algorithm, str), expected) <- Data) {
+      MessageDigest(algorithm).digestString(Source.fromString(str)) shouldBe expected
     }
   }
 
-  test("MessageDigest: InputStream inputs") {
-
-    for (testItem <- Data) {
-      val (algorithm, str) = testItem._1
-      val expected         = testItem._2
-
-      assertResult(expected) {
-        val stream = new ByteArrayInputStream(str.getBytes)
-        MessageDigest(algorithm).digestString(stream)
-      }
+  it should "work with InputStream inputs" in {
+    for (((algorithm, str), expected) <- Data) {
+      val stream = new ByteArrayInputStream(str.getBytes)
+      MessageDigest(algorithm).digestString(stream) shouldBe expected
     }
   }
 }